+++ src/main/java/io/github/astrapi69/tree/ChainableTreeNode.java
@@ -43,7 +43,7 @@ import lombok.experimental.FieldDefaults;
  */
 @Getter
 @Setter
-@EqualsAndHashCode
+@EqualsAndHashCode(exclude = { "child" })
 @ToString(exclude = { "child" })
 @NoArgsConstructor
 @AllArgsConstructor