+++ src/main/java/io/github/astrapi69/tree/LinkedNode.java
@@ -43,7 +43,7 @@ import lombok.experimental.FieldDefaults;
  */
 @Getter
 @Setter
-@EqualsAndHashCode
+@EqualsAndHashCode(exclude = { "next" })
 @ToString(exclude = { "next" })
 @NoArgsConstructor
 @AllArgsConstructor