+++ CHANGELOG.md
@@ -9,6 +9,10 @@ ADDED:
 - new method that get from a given parent the child with a given index
 - new method that get the index from a given child from the given parent
 
+CHANGED:
+
+- added exclude in EqualsAndHashCode for the children field 
+
 Version 5
 -------------